requirements.txt
fastapi
uvicorn
python-multipart
pymongo
qdrant-client

sentence-transformers
PyPDF2
python-dotenv
pydantic
langchain_ollama
